Fix for reboot loop issue in M5ATOMECHO and Audiokit
If a device reboots, memory content of the `mode` member variable would
not be initialized and thus would keep its previous content (mode == MODE_MIC).
And then setReadMode would assume that I2S is already initialized (which is
not true), and then the I2STask would call i2s_read and crash. And this repeats
until a full reset is executed (power off or via EN).

This affects only the mentioned two devices as only these use the `mode`
variable.