Code for #86 Key Press Only Activation Option
Switching to local hotword is now supported (again) in web ui
and also saved in configuration.

Local hotword equals to just using a key press to start Rhasspy listening
on the voice command, no hotword/wakeword has to be given.
Works on all devices which permit support keypress for hotword
(at least AudioKit does)

In local mode, no audio is being streamed until the Rhasspy server
requests voice input.

This required some changes to the handling of audioServer in the I2S task.

A fixed issue is proper detection of broken MQTT connection if
no audio is being streamed. This was not an obvious issue before as almost
always audio was being streamed. The changes to the state machine make
it more robust against MQTTDisconnectedEvent causing trouble when no
wifi connection was established yet.